+++ toolium/driver_wrapper.py
@@ -122,6 +122,7 @@ class DriverWrapper(object):
             # Initialize the config object
             self.config = ExtendedConfigParser.get_config_from_file(prop_filenames)
             self.config_properties_filenames = prop_filenames
+            self.update_magic_config_names()
 
         # Override properties with system properties
         self.config.update_properties(os.environ)
@@ -129,6 +130,20 @@ class DriverWrapper(object):
         # Override properties with behave userdata properties
         if behave_properties:
             self.config.update_properties(behave_properties)
+
+    def update_magic_config_names(self):
+        """Replace '___' for ':' in options names as a workaround of a configparser limitation
+        To set a config property with : in name
+            goog:loggingPrefs = "{'performance': 'ALL', 'browser': 'ALL', 'driver': 'ALL'}"
+        Configure properties.cfg with:
+            goog___loggingPrefs: {'performance': 'ALL', 'browser': 'ALL', 'driver': 'ALL'}
+        """
+        for section in self.config.sections():
+            for option in self.config.options(section):
+                if '___' in option:
+                    option_value = self.config.get(section, option)
+                    self.config.set(section, option.replace('___', ':'), option_value)
+                    self.config.remove_option(section, option)
 
     def configure_visual_baseline(self):
         """Configure baseline directory"""